Add audiovisual project
This commit adds my code for the audiovisual project. As of this
commit, the project is a Windows app that shows an imgui window.
The window presents options for modifying a software synthesizer
comprising a set of configurable oscillators. The app has two threads,
a realtime audio rendering thread and a non-realtime UI thread. The
UI thread and realtime thread communicate via events passed through
lock-free queues. The app supports pan, volume, frequency, and wave
type. It has been checked with address sanitizer.