Fix up the live plotting functionality
After updating the implot library, I missed that there is a new
parameter to PlotLine which was causing the graph to not work. So, I
fixed that in this commit. I also added pausing functionality. I think
there's another issue to deal with later, which is that we're at the
limit of float precision here and the graph looks a bit off at times
with the tiny time step. It's all fine if you load the output wav file
into Audacity. It's still a reasonably useful thing to have so I'm
leaving it for now.